+++ src/trap/_include_trap.jl
@@ -1,0 +1,3 @@
+include("trap.jl")
+include("efield.jl")
+include("bfield.jl")
